+++ standingsrequests/tests/test_integration.py
@@ -1,8 +1,9 @@
 from datetime import timedelta
 from unittest.mock import patch
 
+from django.core.cache import cache
 from django.contrib.auth.models import User
-from django.test import RequestFactory
+from django.test import RequestFactory, TestCase
 from django.urls import reverse
 from django.utils.timezone import now
 
@@ -28,7 +29,7 @@ from ..models import (
 )
 from .. import views
 from .. import tasks
-from ..utils import set_test_logger, NoSocketsTestCase
+from ..utils import set_test_logger
 
 MODULE_PATH_MODELS = "standingsrequests.models"
 MODULE_PATH_MANAGERS = "standingsrequests.managers"
@@ -41,7 +42,7 @@ TEST_SCOPE = "publicData"
 @patch(MODULE_PATH_MODELS + ".STANDINGS_API_CHARID", TEST_STANDINGS_API_CHARID)
 @patch(MODULE_PATH_MANAGERS + ".SR_NOTIFICATIONS_ENABLED", True)
 @patch(MODULE_PATH_MANAGERS + ".STANDINGS_API_CHARID", TEST_STANDINGS_API_CHARID)
-class TestMainUseCases(NoSocketsTestCase):
+class TestMainUseCases(TestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
@@ -81,15 +82,6 @@ class TestMainUseCases(NoSocketsTestCase):
         )
         cls.member_state.member_characters.add(cls.main_2)
         cls.user_manager = User.objects.get(pk=cls.user_manager.pk)
-
-    def setUp(self) -> None:
-        ContactSet.objects.all().delete()
-        self.contact_set = create_contacts_set()
-        StandingRequest.objects.all().delete()
-        StandingRevocation.objects.all().delete()
-        Notification.objects.all().delete()
-        self.member_state.member_characters.add(self.main_1)
-        self.user_requestor = User.objects.get(pk=self.user_requestor.pk)
 
     @patch(MODULE_PATH_TASKS + ".ContactSet.objects.create_new_from_api")
     def _process_standing_requests(self, mock_create_new_from_api):
@@ -136,28 +128,63 @@ class TestMainUseCases(NoSocketsTestCase):
 
     def _remove_standing_for_alt_in_game(self, alt: object) -> None:
         if isinstance(alt, EveCharacter):
-            contact_id = alt.character_id
+            CharacterContact.objects.get(
+                contact_set=self.contact_set, contact_id=alt.character_id
+            ).delete()
         elif isinstance(alt, EveCorporationInfo):
-            contact_id = alt.corporation_id
+            CorporationContact.objects.get(
+                contact_set=self.contact_set, contact_id=alt.corporation_id
+            ).delete()
         else:
             raise NotImplementedError()
-        CharacterContact.objects.get(
-            contact_set=self.contact_set, contact_id=contact_id
-        ).delete()
+
         self.contact_set.refresh_from_db()
 
-    def test_user_requests_standing_for_his_alt_character(self):
+    def _setup_mocks(self, mock_esi_client):
+        mock_Corporation = mock_esi_client.return_value.Corporation
+        mock_Corporation.get_corporations_corporation_id.side_effect = (
+            esi_get_corporations_corporation_id
+        )
+        mock_esi_client.return_value.Universe.post_universe_names.side_effect = (
+            esi_post_universe_names
+        )
+
+    def setUp(self) -> None:
+        cache.clear()
+        ContactSet.objects.all().delete()
+        self.contact_set = create_contacts_set()
+        StandingRequest.objects.all().delete()
+        StandingRevocation.objects.all().delete()
+        Notification.objects.all().delete()
+        self.member_state.member_characters.add(self.main_1)
+        self.user_requestor = User.objects.get(pk=self.user_requestor.pk)
+
+    @patch("standingsrequests.helpers.evecorporation._esi_client", lambda: None)
+    @patch("standingsrequests.helpers.esi_fetch._esi_client")
+    def test_user_requests_standing_for_his_alt_character(self, mock_esi_client):
         """
         given user has permission and user's alt has no standing
         when user requests standing and request is actioned by manager
         then alt has standing and user gets change notification
         """
+        # setup
+        self._setup_mocks(mock_esi_client)
         alt_id = self.alt_1.character_id
 
-        # user requests standing for alt
-        request = self.factory.get(
-            reverse("standingsrequests:request_pilot_standing", args=[alt_id],)
+        # user opens create requests page
+        request = self.factory.get(reverse("standingsrequests:request_entities"))
+        request.user = self.user_requestor
+        response = views.partial_request_entities(request)
+        self.assertEqual(response.status_code, 200)
+
+        # make sure link for requesting standing is shown to user
+        request_standing_url = reverse(
+            "standingsrequests:request_pilot_standing", args=[alt_id],
         )
+        self.assertIn(request_standing_url, response.content.decode("utf8"))
+
+        # user requests standing for alt
+        request = self.factory.get(request_standing_url)
         request.user = self.user_requestor
         response = views.request_pilot_standing(request, alt_id)
         self.assertEqual(response.status_code, 302)
@@ -166,9 +193,8 @@ class TestMainUseCases(NoSocketsTestCase):
         self.assertFalse(my_request.is_effective)
         self.assertEqual(my_request.user, self.user_requestor)
 
+        # set standing in game and mark as actioned
         self._set_standing_for_alt_in_game(self.alt_1)
-
-        # mark standing as actioned
         request = self.factory.put(
             reverse("standingsrequests:manage_requests_write", args=[alt_id],)
         )
@@ -188,12 +214,67 @@ class TestMainUseCases(NoSocketsTestCase):
         self.assertIsNotNone(my_request.effective_date)
         self.assertTrue(Notification.objects.filter(user=self.user_requestor).exists())
 
-    @patch("standingsrequests.helpers.evecorporation.cache")
-    @patch("standingsrequests.helpers.esi_fetch._esi_client")
     @patch("standingsrequests.helpers.evecorporation._esi_client", lambda: None)
-    def test_user_requests_standing_for_his_alt_corporation(
-        self, mock_esi_client, mock_cache
-    ):
+    @patch("standingsrequests.helpers.esi_fetch._esi_client")
+    def test_user_requests_revocation_for_his_alt_character(self, mock_esi_client):
+        """
+        given user's alt has standing and user has permission
+        when user requests revocation and request is actioned by manager
+        then alt's standing is removed and user gets change notification
+        """
+
+        # setup
+        self._setup_mocks(mock_esi_client)
+        alt_id = self.alt_1.character_id
+        self._set_standing_for_alt_in_game(self.alt_1)
+        my_request = self._create_standing_for_alt(self.alt_1)
+
+        # user opens create requests page
+        request = self.factory.get(reverse("standingsrequests:request_entities"))
+        request.user = self.user_requestor
+        response = views.partial_request_entities(request)
+        self.assertEqual(response.status_code, 200)
+
+        # make sure link for revoking standing is shown to user
+        revoke_standing_url = reverse(
+            "standingsrequests:remove_pilot_standing", args=[alt_id],
+        )
+        self.assertIn(revoke_standing_url, response.content.decode("utf8"))
+
+        # user requests revocation for alt
+        request = self.factory.get(revoke_standing_url)
+        request.user = self.user_requestor
+        response = views.remove_pilot_standing(request, alt_id)
+        self.assertEqual(response.status_code, 302)
+        self.assertTrue(my_request.is_effective)
+        self.assertEqual(my_request.user, self.user_requestor)
+        my_revocation = StandingRevocation.objects.get(contact_id=alt_id)
+        self.assertFalse(my_revocation.is_effective)
+
+        # remove standing for alt in game and mark as actioned
+        self._remove_standing_for_alt_in_game(self.alt_1)
+
+        request = self.factory.put(
+            reverse("standingsrequests:manage_revocations_write", args=[alt_id],)
+        )
+        request.user = self.user_manager
+        response = views.manage_revocations_write(request, alt_id)
+        self.assertEqual(response.status_code, 204)
+        my_revocation.refresh_from_db()
+        self.assertEqual(my_revocation.action_by, self.user_manager)
+        self.assertIsNotNone(my_revocation.action_date)
+        self.assertFalse(my_revocation.is_effective)
+
+        self._process_standing_requests()
+
+        # validate results
+        self.assertFalse(StandingRequest.objects.filter(contact_id=alt_id).exists())
+        self.assertFalse(StandingRevocation.objects.filter(contact_id=alt_id).exists())
+        self.assertTrue(Notification.objects.filter(user=self.user_requestor).exists())
+
+    @patch("standingsrequests.helpers.evecorporation._esi_client", lambda: None)
+    @patch("standingsrequests.helpers.esi_fetch._esi_client")
+    def test_user_requests_standing_for_his_alt_corporation(self, mock_esi_client):
         """
         given user has permission and user's alt has no standing
         and all corporation members have tokens
@@ -202,23 +283,26 @@ class TestMainUseCases(NoSocketsTestCase):
         """
 
         # setup
-        mock_Corporation = mock_esi_client.return_value.Corporation
-        mock_Corporation.get_corporations_corporation_id.side_effect = (
-            esi_get_corporations_corporation_id
-        )
-        mock_esi_client.return_value.Universe.post_universe_names.side_effect = (
-            esi_post_universe_names
-        )
-        mock_cache.get.return_value = None
+        self._setup_mocks(mock_esi_client)
         alt_id = self.alt_corporation.corporation_id
         add_character_to_user(
             self.user_requestor, self.alt_2, scopes=[TEST_SCOPE],
         )
 
-        # user requests standing for alt
-        request = self.factory.get(
-            reverse("standingsrequests:request_corp_standing", args=[alt_id],)
+        # user opens create requests page
+        request = self.factory.get(reverse("standingsrequests:request_entities"))
+        request.user = self.user_requestor
+        response = views.partial_request_entities(request)
+        self.assertEqual(response.status_code, 200)
+
+        # make sure link for requesting standing is shown to user
+        request_standing_url = reverse(
+            "standingsrequests:request_corp_standing", args=[alt_id],
         )
+        self.assertIn(request_standing_url, response.content.decode("utf8"))
+
+        # user requests standing for alt
+        request = self.factory.get(request_standing_url)
         request.user = self.user_requestor
         response = views.request_corp_standing(request, alt_id)
         self.assertEqual(response.status_code, 302)
@@ -227,9 +311,8 @@ class TestMainUseCases(NoSocketsTestCase):
         self.assertFalse(my_request.is_effective)
         self.assertEqual(my_request.user, self.user_requestor)
 
+        # set standing in game and mark as actioned
         self._set_standing_for_alt_in_game(self.alt_corporation)
-
-        # mark standing as actioned
         request = self.factory.put(
             reverse("standingsrequests:manage_requests_write", args=[alt_id],)
         )
@@ -247,9 +330,11 @@ class TestMainUseCases(NoSocketsTestCase):
         my_request.refresh_from_db()
         self.assertTrue(my_request.is_effective)
         self.assertIsNotNone(my_request.effective_date)
-        # self.assertTrue(Notification.objects.filter(user=self.user_requestor).exists())
+        self.assertTrue(Notification.objects.filter(user=self.user_requestor).exists())
 
-    def test_user_requests_revocation_for_his_alt_character(self):
+    @patch("standingsrequests.helpers.evecorporation._esi_client", lambda: None)
+    @patch("standingsrequests.helpers.esi_fetch._esi_client")
+    def test_user_requests_revocation_for_his_alt_corporation(self, mock_esi_client):
         """
         given user's alt has standing and user has permission
         when user requests revocation and request is actioned by manager
@@ -257,25 +342,35 @@ class TestMainUseCases(NoSocketsTestCase):
         """
 
         # setup
-        alt_id = self.alt_1.character_id
-        self._set_standing_for_alt_in_game(self.alt_1)
-        my_request = self._create_standing_for_alt(self.alt_1)
+        self._setup_mocks(mock_esi_client)
+        alt_id = self.alt_corporation.corporation_id
+        self._set_standing_for_alt_in_game(self.alt_corporation)
+        my_request = self._create_standing_for_alt(self.alt_corporation)
+
+        # user opens create requests page
+        request = self.factory.get(reverse("standingsrequests:request_entities"))
+        request.user = self.user_requestor
+        response = views.partial_request_entities(request)
+        self.assertEqual(response.status_code, 200)
+
+        # make sure link for revoking standing is shown to user
+        revoke_standing_url = reverse(
+            "standingsrequests:remove_corp_standing", args=[alt_id],
+        )
+        self.assertIn(revoke_standing_url, response.content.decode("utf8"))
 
         # user requests revocation for alt
-        request = self.factory.get(
-            reverse("standingsrequests:remove_pilot_standing", args=[alt_id],)
-        )
+        request = self.factory.get(revoke_standing_url)
         request.user = self.user_requestor
-        response = views.remove_pilot_standing(request, alt_id)
+        response = views.remove_corp_standing(request, alt_id)
         self.assertEqual(response.status_code, 302)
         self.assertTrue(my_request.is_effective)
         self.assertEqual(my_request.user, self.user_requestor)
         my_revocation = StandingRevocation.objects.get(contact_id=alt_id)
         self.assertFalse(my_revocation.is_effective)
 
-        self._remove_standing_for_alt_in_game(self.alt_1)
-
-        # mark revocation as actioned
+        # remove standing for alt in game and mark as actioned
+        self._remove_standing_for_alt_in_game(self.alt_corporation)
         request = self.factory.put(
             reverse("standingsrequests:manage_revocations_write", args=[alt_id],)
         )
@@ -301,7 +396,7 @@ class TestMainUseCases(NoSocketsTestCase):
         then alt's standing is removed and user gets change notification
         """
 
-        # Setup
+        # setup
         alt_id = self.alt_1.character_id
         self._create_standing_for_alt(self.alt_1)